refactor: clear the clang-tidy Phase-0 baseline + enforce the gate
Fixes all ~30 findings from the report-only Phase-0 set and flips the CI
`clang-tidy` job from report-only to enforcing, so the cleaned set stays clean.

Highlights (behaviour-preserving throughout):
- performance-unnecessary-value-param / -copy-initialization: pass/bind by
  `const&` where the body only reads.
- performance-no-automatic-move: drop needless `const` on returned locals so
  NRVO/move kicks in.
- clang-analyzer VirtualCall: qualify `stop()` -> `ClassName::stop()` in
  destructors (the virtual-call-from-dtor footgun; static dispatch is what we
  actually want there).
- bugprone-implicit-widening-of-multiplication-result: widen an operand before
  the multiply in a few config/TTL/bench spots. Defensive only — the operands
  are small constants today, but now they won't bite if inputs ever grow.
- bugprone-branch-clone / switch-missing-default-case / override-with-different-
  visibility: merged identical arms, added `default: break;`, and made
  `MainWindow::eventFilter` public to match QObject.
- ProcessResolverFactory: inlined the feature gates so they don't read as dead
  stores in the no-attribution build config.

6 narrow NOLINTs remain, each with a one-line reason: 4x Qt's mandated
QDBusArgument extraction-operator signature (can't change it), and 1x a DNS
copy that must outlive an iterator `erase` (const& there would be a UAF).

`bugprone-narrowing-conversions` and the modernize/readability families stay
parked for later phases. 0 findings now; 36/36 tests pass.

+++ src/agent/Config.cpp
@@ -97,7 +97,7 @@ IdleManager::Config loadIdleConfig(const QString &path)
     // multiplication in 32-bit on the raw INI value (which a typo could
     // easily make huge) is UB on overflow — the previous code clamped after
     // multiplying, which doesn't save you from the overflow itself.
-    constexpr qint64 kMaxSec = 24 * 60 * 60;
+    constexpr qint64 kMaxSec = qint64(24) * 60 * 60;
     auto secsToMs = [&](const char *key, const QString &iniKey, int defaultMs) -> int {
         const qint64 raw = ini.value(iniKey, qint64(defaultMs) / 1000).toLongLong();
         if (raw < 0 || raw > kMaxSec) {
